+++ src/backend/imap/backend.rs
@@ -2,11 +2,19 @@
 //!
 //! This module contains the definition of the IMAP backend.
 
-use imap::types::NameAttribute;
+use imap::{extensions::idle::SetReadTimeout, types::NameAttribute};
 use log::{debug, log_enabled, trace, Level};
 use native_tls::{TlsConnector, TlsStream};
-use std::{any::Any, collections::HashSet, convert::TryInto, net::TcpStream, result, thread};
+use std::{
+    any::Any,
+    collections::HashSet,
+    convert::TryInto,
+    io::{self, Read, Write},
+    net::TcpStream,
+    result, thread,
+};
 use thiserror::Error;
+use utf7_imap::{decode_utf7_imap as decode_utf7, encode_utf7_imap as encode_utf7};
 
 use crate::{
     account, backend, email, envelope, flag, process, AccountConfig, Backend, Email, Envelopes,
@@ -96,12 +104,51 @@ pub enum Error {
 
 pub type Result<T> = result::Result<T, Error>;
 
-type ImapSess = imap::Session<TlsStream<TcpStream>>;
+enum ImapSessionStream {
+    Tls(TlsStream<TcpStream>),
+    Tcp(TcpStream),
+}
+
+impl SetReadTimeout for ImapSessionStream {
+    fn set_read_timeout(&mut self, timeout: Option<std::time::Duration>) -> imap::Result<()> {
+        match self {
+            Self::Tls(stream) => stream.set_read_timeout(timeout),
+            Self::Tcp(stream) => stream.set_read_timeout(timeout),
+        }
+    }
+}
+
+impl Read for ImapSessionStream {
+    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
+        match self {
+            Self::Tls(stream) => stream.read(buf),
+            Self::Tcp(stream) => stream.read(buf),
+        }
+    }
+}
+
+impl Write for ImapSessionStream {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        match self {
+            Self::Tls(stream) => stream.write(buf),
+            Self::Tcp(stream) => stream.write(buf),
+        }
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        match self {
+            Self::Tls(stream) => stream.flush(),
+            Self::Tcp(stream) => stream.flush(),
+        }
+    }
+}
+
+type ImapSession = imap::Session<ImapSessionStream>;
 
 pub struct ImapBackend<'a> {
     account_config: &'a AccountConfig,
     imap_config: &'a ImapConfig,
-    sess: Option<ImapSess>,
+    sess: Option<ImapSession>,
 }
 
 impl<'a> ImapBackend<'a> {
@@ -113,7 +160,7 @@ impl<'a> ImapBackend<'a> {
         }
     }
 
-    fn sess(&mut self) -> Result<&mut ImapSess> {
+    fn sess(&mut self) -> Result<&mut ImapSession> {
         if self.sess.is_none() {
             debug!("create TLS builder");
             debug!("insecure: {}", self.imap_config.insecure());
@@ -132,12 +179,17 @@ impl<'a> ImapBackend<'a> {
             if self.imap_config.starttls() {
                 client_builder.starttls();
             }
-            let client = client_builder
-                .connect(|domain, tcp| Ok(TlsConnector::connect(&builder, domain, tcp)?))
-                .map_err(Error::ConnectImapServerError)?;
 
-            debug!("create session");
-            debug!("login: {}", self.imap_config.login);
+            let client = if self.imap_config.ssl() {
+                client_builder.connect(|domain, tcp| {
+                    let connector = TlsConnector::connect(&builder, domain, tcp)?;
+                    Ok(ImapSessionStream::Tls(connector))
+                })
+            } else {
+                client_builder.connect(|_, tcp| Ok(ImapSessionStream::Tcp(tcp)))
+            }
+            .map_err(Error::ConnectImapServerError)?;
+
             let mut sess = client
                 .login(&self.imap_config.login, &self.imap_config.passwd()?)
                 .map_err(|res| Error::LoginImapServerError(res.0))?;
@@ -259,10 +311,9 @@ impl<'a> ImapBackend<'a> {
             let cmds = self.imap_config.watch_cmds().clone();
             thread::spawn(move || {
                 debug!("batch execution of {} cmd(s)", cmds.len());
-                cmds.iter().for_each(|cmd| {
-                    debug!("running command {:?}…", cmd);
-                    let res = process::run(cmd);
-                    debug!("{:?}", res);
+                cmds.iter().for_each(|cmd| match process::run(cmd, &[]) {
+                    Err(_) => (),
+                    Ok(_) => (),
                 })
             });
 
@@ -272,20 +323,17 @@ impl<'a> ImapBackend<'a> {
 }
 
 impl<'a> Backend<'a> for ImapBackend<'a> {
-    fn folder_add(&mut self, mbox: &str) -> backend::Result<()> {
-        trace!(">> add folder");
+    fn folder_add(&mut self, folder: &str) -> backend::Result<()> {
+        let folder = encode_utf7(folder.to_owned());
 
         self.sess()?
-            .create(mbox)
-            .map_err(|err| Error::CreateMboxError(err, mbox.to_owned()))?;
+            .create(&folder)
+            .map_err(|err| Error::CreateMboxError(err, folder.to_owned()))?;
 
-        trace!("<< add folder");
         Ok(())
     }
 
     fn folder_list(&mut self) -> backend::Result<Folders> {
-        trace!(">> get imap folders");
-
         let imap_mboxes = self
             .sess()?
             .list(Some(""), Some("*"))
@@ -295,7 +343,7 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
                 .iter()
                 .map(|imap_mbox| Folder {
                     delim: imap_mbox.delimiter().unwrap_or_default().into(),
-                    name: imap_mbox.name().into(),
+                    name: decode_utf7(imap_mbox.name().into()),
                     desc: imap_mbox
                         .attributes()
                         .iter()
@@ -317,27 +365,27 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
         Ok(mboxes)
     }
 
-    fn folder_delete(&mut self, mbox: &str) -> backend::Result<()> {
-        trace!(">> delete imap folder");
+    fn folder_delete(&mut self, folder: &str) -> backend::Result<()> {
+        let folder = encode_utf7(folder.to_owned());
 
         self.sess()?
-            .delete(mbox)
-            .map_err(|err| Error::DeleteMboxError(err, mbox.to_owned()))?;
+            .delete(&folder)
+            .map_err(|err| Error::DeleteMboxError(err, folder.to_owned()))?;
 
-        trace!("<< delete imap folder");
         Ok(())
     }
 
     fn envelope_list(
         &mut self,
-        mbox: &str,
+        folder: &str,
         page_size: usize,
         page: usize,
     ) -> backend::Result<Envelopes> {
+        let folder = encode_utf7(folder.to_owned());
         let last_seq = self
             .sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?
             .exists as usize;
         debug!("last sequence number: {:?}", last_seq);
         if last_seq == 0 {
@@ -365,16 +413,17 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
 
     fn envelope_search(
         &mut self,
-        mbox: &str,
+        folder: &str,
         query: &str,
         sort: &str,
         page_size: usize,
         page: usize,
     ) -> backend::Result<Envelopes> {
+        let folder = encode_utf7(folder.to_owned());
         let last_seq = self
             .sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?
             .exists;
         debug!("last sequence number: {:?}", last_seq);
         if last_seq == 0 {
@@ -386,7 +435,7 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
         let seqs: Vec<String> = if sort.is_empty() {
             self.sess()?
                 .search(query)
-                .map_err(|err| Error::SearchMsgsError(err, mbox.to_owned(), query.to_owned()))?
+                .map_err(|err| Error::SearchMsgsError(err, folder.to_owned(), query.to_owned()))?
                 .iter()
                 .map(|seq| seq.to_string())
                 .collect()
@@ -394,7 +443,7 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
             let sort: envelope::imap::SortCriteria = sort.try_into()?;
             self.sess()?
                 .sort(&sort, imap::extensions::sort::SortCharset::Utf8, query)
-                .map_err(|err| Error::SortMsgsError(err, mbox.to_owned(), query.to_owned()))?
+                .map_err(|err| Error::SortMsgsError(err, folder.to_owned(), query.to_owned()))?
                 .iter()
                 .map(|seq| seq.to_string())
                 .collect()
@@ -413,25 +462,27 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
         Ok(envelopes)
     }
 
-    fn email_add(&mut self, mbox: &str, msg: &[u8], flags: &str) -> backend::Result<String> {
+    fn email_add(&mut self, folder: &str, msg: &[u8], flags: &str) -> backend::Result<String> {
+        let folder = encode_utf7(folder.to_owned());
         let flags: Flags = flags.into();
         self.sess()?
-            .append(mbox, msg)
+            .append(&folder, msg)
             .flags(flag::imap::into_raws(&flags))
             .finish()
-            .map_err(|err| Error::AppendMsgError(err, mbox.to_owned()))?;
+            .map_err(|err| Error::AppendMsgError(err, folder.to_owned()))?;
         let last_seq = self
             .sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?
             .exists;
         Ok(last_seq.to_string())
     }
 
-    fn email_get(&mut self, mbox: &str, seq: &str) -> backend::Result<Email> {
+    fn email_get(&mut self, folder: &str, seq: &str) -> backend::Result<Email> {
+        let folder = encode_utf7(folder.to_owned());
         self.sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?;
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?;
         let fetches = self
             .sess()?
             .fetch(seq, "(FLAGS INTERNALDATE BODY[])")
@@ -449,53 +500,56 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
         Ok(msg)
     }
 
-    fn email_copy(&mut self, mbox_src: &str, mbox_dst: &str, seq: &str) -> backend::Result<()> {
-        let msg = self.email_get(&mbox_src, seq)?.raw;
-        self.email_add(&mbox_dst, &msg, "seen")?;
+    fn email_copy(&mut self, folder: &str, folder_target: &str, seq: &str) -> backend::Result<()> {
+        let email = self.email_get(folder, seq)?.raw;
+        self.email_add(folder_target, &email, "seen")?;
         Ok(())
     }
 
-    fn email_move(&mut self, mbox_src: &str, mbox_dst: &str, seq: &str) -> backend::Result<()> {
-        let msg = self.email_get(mbox_src, seq)?.raw;
-        self.flags_add(mbox_src, seq, "seen deleted")?;
-        self.email_add(&mbox_dst, &msg, "seen")?;
+    fn email_move(&mut self, folder: &str, folder_target: &str, seq: &str) -> backend::Result<()> {
+        let msg = self.email_get(folder, seq)?.raw;
+        self.flags_add(folder, seq, "seen deleted")?;
+        self.email_add(folder_target, &msg, "seen")?;
         Ok(())
     }
 
-    fn email_delete(&mut self, mbox: &str, seq: &str) -> backend::Result<()> {
-        self.flags_add(mbox, seq, "deleted")
+    fn email_delete(&mut self, folder: &str, seq: &str) -> backend::Result<()> {
+        self.flags_add(folder, seq, "deleted")
     }
 
-    fn flags_add(&mut self, mbox: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+    fn flags_add(&mut self, folder: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+        let folder = encode_utf7(folder.to_owned());
         let flags: Flags = flags.into();
         self.sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?;
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?;
         self.sess()?
             .store(seq_range, format!("+FLAGS ({})", flags))
             .map_err(|err| Error::AddFlagsError(err, flags.to_owned(), seq_range.to_owned()))?;
         self.sess()?
             .expunge()
-            .map_err(|err| Error::ExpungeError(err, mbox.to_owned()))?;
+            .map_err(|err| Error::ExpungeError(err, folder.to_owned()))?;
         Ok(())
     }
 
-    fn flags_set(&mut self, mbox: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+    fn flags_set(&mut self, folder: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+        let folder = encode_utf7(folder.to_owned());
         let flags: Flags = flags.into();
         self.sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?;
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?;
         self.sess()?
             .store(seq_range, format!("FLAGS ({})", flags))
             .map_err(|err| Error::SetFlagsError(err, flags.to_owned(), seq_range.to_owned()))?;
         Ok(())
     }
 
-    fn flags_delete(&mut self, mbox: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+    fn flags_delete(&mut self, folder: &str, seq_range: &str, flags: &str) -> backend::Result<()> {
+        let folder = encode_utf7(folder.to_owned());
         let flags: Flags = flags.into();
         self.sess()?
-            .select(mbox)
-            .map_err(|err| Error::SelectMboxError(err, mbox.to_owned()))?;
+            .select(&folder)
+            .map_err(|err| Error::SelectMboxError(err, folder.to_owned()))?;
         self.sess()?
             .store(seq_range, format!("-FLAGS ({})", flags))
             .map_err(|err| Error::DelFlagsError(err, flags.to_owned(), seq_range.to_owned()))?;
@@ -503,8 +557,6 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
     }
 
     fn disconnect(&mut self) -> backend::Result<()> {
-        trace!(">> imap logout");
-
         if let Some(ref mut sess) = self.sess {
             debug!("logout from imap server");
             sess.logout().map_err(Error::LogoutError)?;
@@ -512,7 +564,6 @@ impl<'a> Backend<'a> for ImapBackend<'a> {
             debug!("no session found");
         }
 
-        trace!("<< imap logout");
         Ok(())
     }